Fixed namespace package issue between ensembl-production and ensembl-metadata-api (same base path ensembl.production.metadata) Moved the one from production to `ensembl.production.factories/metadata`
